pkg/dora/{dora,parse,query}: change the dora query layer to use a []byte
Change query to be a []byte and update a few places where string <-> byte conversions happen to use the new danger package. As far as I can tell these scenarios should be safe for now. Commented out tests are a skeleton to start building out the Set<T> functionality against.
Change-Id: I1130b3ad84cd991079535508c324cffdd4f8a223